api/requirements.txt
Django==3.2.15
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
ipython==7.34.0
